Move Seq element implicit derivations to the trait's companion object
Fixes #252
