+++ core/shared/src/main/scala/me/shadaj/scalapy/py/package.scala
@@ -39,26 +39,6 @@ package object py extends PyMacros {
     }
   }
 
-  trait ConvertableToSeqElem[T] {
-    def convertCopy(v: T): Platform.Pointer
-    def convertProxy(v: T): PyValue
-  }
-
-  implicit def seqConvertableSeqElem[T, C](implicit ev: C => scala.collection.Seq[T], elemConvertable: ConvertableToSeqElem[T]): ConvertableToSeqElem[C] = new ConvertableToSeqElem[C] {
-    def convertCopy(v: C): Platform.Pointer = {
-      CPythonInterpreter.createListCopy(ev(v), elemConvertable.convertCopy)
-    }
-
-    def convertProxy(v: C): PyValue = {
-      CPythonInterpreter.createListProxy(ev(v), elemConvertable.convertProxy)
-    }
-  }
-
-  implicit def writableSeqElem[T](implicit writer: Writer[T]): ConvertableToSeqElem[T] = new ConvertableToSeqElem[T] {
-    def convertCopy(v: T): Platform.Pointer = writer.writeNative(v)
-    def convertProxy(v: T): PyValue = writer.write(v)
-  }
-
   implicit class SeqConverters[T, C](seq: C)(implicit ev: C => scala.collection.Seq[T]) {
     def toPythonCopy(implicit elemWriter: ConvertableToSeqElem[T]): Any = {
       Any.populateWith(PyValue.fromNew(implicitly[ConvertableToSeqElem[C]].convertCopy(seq)))
